backend/requirement.txt
fastapi
uvicorn
easyocr
pillow
numpy
transformers
torch
python-multipart
